requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
--find-links https://data.pyg.org/whl/torch-2.8.0+cu126.html

torch==2.8.0
torchvision
torch_scatter
torch_sparse

opencv-python-headless
mesh2sdf
smplx
pymeshlab
numpy==1.23.5
diffusers==0.27.2
transformers==4.39.3
huggingface_hub==0.24.6
imageio
imageio-ffmpeg
open3d
iopath
tensorboard
trimesh
pytorch_lightning==1.9.3
matplotlib
easydict
cholespy
fire
omegaconf==2.1.1
einops
accelerate
kornia
xformers
chumpy
polyscope
scikit-learn
seaborn
potpourri3d
tqdm
human-body-prior
joblib
ema_pytorch
robust-laplacian
ipykernel
libigl